S2.0b: convert_model.py + Qwen byte-identical round-trip PROVEN
convert_model.py loads the calibrated QuantizedModel, runs the C24/C25/C27 folds
OFFLINE (float never reaches runtime), emits all 8 model sections via sslm_model_writer.
sslm_inspect extended to parse every section through the ModelView (the read-back proof).

Round-trip on real Qwen2.5-1.5B (1.585 GB .sslm): C++ loader OK, fingerprint byte-identical
(cd6a0e0b...), MODELVIEW OK — every section parsed, values match Python source (Config dims,
composition-constant/kv-landing tuples, tensor shapes).

Empirical fix caught at source: dynamic-bias codes are int64 (~10^14 at q_b=30), not int32
as the plan text implied — Biases dtype Int32->Int64 (loader ExpectedDtype C++/Python, docs,
+ two S0 fixtures updated to match). Suite 597/0.

+++ src/artifact.cpp
@@ -48,7 +48,8 @@ bool IsKnownSectionType(uint32_t type) noexcept {
 SslmDtype ExpectedDtype(uint32_t type) noexcept {
 	switch (static_cast<SslmSectionType>(type)) {
 		case SslmSectionType::Weights: return SslmDtype::Int8;
-		case SslmSectionType::Biases: return SslmDtype::Int32;
+		// Biases carries the C28 dynamic-bias codes (int64 — they reach ~10^14 at q_b=30).
+		case SslmSectionType::Biases: return SslmDtype::Int64;
 		// WeightScales is a WSC1 tensor manifest of int32 (identity,mult,shift) fold ops.
 		case SslmSectionType::WeightScales: return SslmDtype::Int32;
 		case SslmSectionType::RopeTables: return SslmDtype::Int64;
